src/format: use u32 and std::process::id not libc::getpid
Since 0.3.2, libc::getpid is used to get the process id and store the
provided id in Formatters. This has the marked side effect of being
unsafe and adding an unnecessary crate dependency. As such, this change
moves to using std::process::id, which likely was not stable at the
time, for better stdlib integration.

Unfortunately, this is a breaking change, since the current i32s are
part of the external interface, but could be implemented as compatible
if the extant symbols were simply deprecated and new ones added.